+++ DestinyEngine/src/Destiny/Events/WindowEvent.hpp
@@ -14,8 +14,8 @@ namespace Destiny
 		WindowResizeEvent(unsigned int width, unsigned int height)
 			: m_Width(width), m_Height(height) {}
 
-		unsigned int getWidth() const { return m_Width; }
-		unsigned int getHeight() const { return m_Height; }
+		uint16_t getWidth() const { return m_Width; }
+		uint16_t getHeight() const { return m_Height; }
 
 		std::string toString() const override
 		{
@@ -25,9 +25,9 @@ namespace Destiny
 		}
 
 		EVENT_CLASS_TYPE(WindowResize)
-			EVENT_CLASS_CATEGORY(EventCategoryWindow)
+		EVENT_CLASS_CATEGORY(EventCategoryWindow)
 	private:
-		unsigned int m_Width, m_Height;
+		uint16_t m_Width, m_Height;
 	};
 
 	class WindowCloseEvent : public Event
